+++ .gitignore
@@ -34,6 +34,8 @@ MANIFEST
 *.manifest
 *.spec
 
+
+
 # Installer logs
 pip-log.txt
 pip-delete-this-directory.txt
@@ -57,10 +59,12 @@ cover/
 *.mo
 *.pot
 
+# SQLITE
+*.db
+
 # Django stuff:
 *.log
 local_settings.py
-*.db
 db.sqlite3
 db.sqlite3-journal
