+++ system_atm/config.py
@@ -1,0 +1,5 @@
+from os import system, name
+from sqlite3 import Connection, connect, Cursor
+
+
+
